updating the notesHelper
